Refactor to allow for styling of HTML export
