+++ shared/editor/components/Styles.ts
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+import style, { Props } from "../../styles/editor";
+
+const EditorContainer = styled.div<Props>`
+  ${style};
+`;
+
+export default EditorContainer;
