+++ shared/styles/globals.ts
@@ -1,6 +1,6 @@
 import { createGlobalStyle } from "styled-components";
 import styledNormalize from "styled-normalize";
-import { breakpoints, depths } from "@shared/styles";
+import { breakpoints, depths } from ".";
 
 export default createGlobalStyle`
   ${styledNormalize}
